Added Red and Blue masking to MergeViewer
This code adds proper masking for this years game. While running MergeViewer you can press space to change between red and blue masks.

I will be updating the MergeFRCPipeline half of it later, but I am waiting to test it on a pi first before commiting.

Code was reviewed to be good by my dad.